requirements.txt
pandas>=2.3.0
scispacy>=0.5.0
spacy>=3.7.0
transformers>=4.56.0
faiss-cpu>=1.12.0
networkx>=3.5.0
openai>=1.107.0
torch>=2.8.0
matplotlib>=3.10.0
seaborn>=0.13.0
numpy>=1.26.0
scikit-learn>=1.7.0
tqdm>=4.67.0
streamlit>=1.49.0
plotly>=6.3.0
pyvis>=0.3.2
python-dotenv>=1.1.0
jupyter>=1.1.0
